+++ form/src/test/java/com/ritense/form/BaseTest.java
@@ -19,8 +19,8 @@ package com.ritense.form;
 import com.ritense.document.domain.impl.JsonSchema;
 import com.ritense.document.domain.impl.JsonSchemaDocumentDefinition;
 import com.ritense.document.domain.impl.JsonSchemaDocumentDefinitionId;
-import com.ritense.form.config.SpringContextHelper;
 import com.ritense.form.domain.FormIoFormDefinition;
+import com.ritense.form.domain.SpringContextHelper;
 import com.ritense.valtimo.contract.form.FormFieldDataResolver;
 import org.apache.commons.io.IOUtils;
 import org.springframework.context.ApplicationContext;
